fix(Skips/WinShardEffectSkip): update patch for Haste 1.3.b
Haste 1.3.b changes `RunHandler.CompleteRun()` a bit. It moves the calls to `WinShardEffect.PlayEffect()` and `MonoFunctions.DelayCall(RunHandler.TransitionToPostGame)` to an inline method which is passed to `MonoFunctions.DelayCall()`. This seems to be so the animation can start a few seconds later if need be.

To fix the SkippableSkip hook, we now find and hook into that inline method instead of `RunHandler.CompleteRun()` directly.